Tidied up Sidebar component and slugify util method a bit
