+++ www/src/components/Sidebar/Sidebar.js
@@ -1,28 +1,19 @@
 import React from 'react';
-import {getDefaultActiveSection} from './utils';
 import Section from './Section';
 import styles from './Sidebar.module.scss';
-
-// TODO (HACK) This data should be passed in as a parameter
-import sectionList from '../../../../docs/_data/nav_docs.yml';
-
-// TODO Support sticky behavior
 
 class Sidebar extends React.Component {
   constructor(props, context) {
     super(props, context);
 
     this.state = {
-      activeSection: getDefaultActiveSection(
-        props.location.pathname,
-        sectionList,
-      ),
+      activeSection: props.defaultActiveSection,
     };
   }
 
   render() {
-    const activeSection = this.state.activeSection;
-    const pathname = this.props.location.pathname;
+    const {location, sectionList} = this.props;
+    const {activeSection} = this.state;
 
     return (
       <nav className={styles.Sidebar}>
@@ -31,7 +22,7 @@ class Sidebar extends React.Component {
             isActive={activeSection === section}
             key={index}
             onClick={() => this._toggleSection(section)}
-            pathname={pathname}
+            pathname={location.pathname}
             section={section}
           />
         ))}
