+++ www/src/utils/slugify.js
@@ -1,0 +1,3 @@
+import slugify from 'underscore.string/slugify';
+
+export default (string) => slugify(string) + '.html';
